+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Auto crop could fail in certain circumstances.
 - Bug when working with VC-1 files and generating images due to auto crop and basic comparison mode.
 - Media search bug when stripping text is fixed.
+- Bug for template validation when it comes to Unit3D trackers.
 
 ### Removed